address thermos quality findings
