+++ packages/core/src/project-info/discover-project.ts
@@ -16,17 +16,14 @@ import {
   findShopifyFlashListVersion,
   SHOPIFY_FLASH_LIST_PACKAGE_NAME,
 } from "./find-shopify-flash-list-version.js";
+import { resolveCatalogBackedDependencyVersion } from "./resolve-catalog-backed-dependency-version.js";
 import { getPreactVersion } from "./get-preact-version.js";
 import { hasTanStackQuery } from "./has-tanstack-query.js";
 import { someWorkspacePackageJson } from "./some-workspace-package-json.js";
 import { isPackageJsonReanimatedAware } from "./utils/is-package-json-reanimated-aware.js";
 import { readPackageJson } from "./read-package-json.js";
 import { getLowestDependencyMajor } from "./utils/dependency-version-spec.js";
-import {
-  extractCatalogName,
-  isCatalogReference,
-  resolveCatalogVersion,
-} from "./resolve-catalog-version.js";
+import { isCatalogReference, resolveCatalogVersion } from "./resolve-catalog-version.js";
 import { parseReactMajor } from "./parse-react-major.js";
 import { parseZodMajor } from "./parse-zod-major.js";
 import { resolveEffectiveReactMajor } from "./resolve-effective-react-major.js";
@@ -196,64 +193,23 @@ export const discoverProject = (directory: string): ProjectInfo => {
     framework === "react-native" ||
     hasReactNativeWorkspaceAnywhere(directory, packageJson);
 
-  // Expo implies React Native, so a project that isn't RN-aware anywhere
-  // can never be Expo — skip the (root + workspace) `expo` version lookup
-  // entirely in that case.
-  let expoVersion = hasReactNativeWorkspace ? findExpoVersion(directory, packageJson) : null;
-
-  // `findExpoVersion` returns the raw `expo` spec, which can be a `catalog:`
-  // reference in pnpm-catalog monorepos. Resolve it the same way `react` /
-  // `tailwind` / `zod` are resolved above, so the Expo SDK major can be
-  // parsed — an unresolved `catalog:` spec would leave `expoVersion` non-null
-  // (Expo checks still run) yet leave the SDK unresolvable, silently disabling
-  // every SDK-gated Expo rule.
-  if (expoVersion !== null && isCatalogReference(expoVersion)) {
-    const catalogName = extractCatalogName(expoVersion);
-    let resolvedExpoVersion = resolveCatalogVersion(packageJson, "expo", directory, catalogName);
-    if (!resolvedExpoVersion) {
-      const monorepoRoot = findMonorepoRoot(directory);
-      if (monorepoRoot) {
-        const monorepoPackageJsonPath = path.join(monorepoRoot, "package.json");
-        if (isFile(monorepoPackageJsonPath)) {
-          resolvedExpoVersion = resolveCatalogVersion(
-            readPackageJson(monorepoPackageJsonPath),
-            "expo",
-            monorepoRoot,
-            catalogName,
-          );
-        }
-      }
-    }
-    expoVersion = resolvedExpoVersion ?? expoVersion;
-  }
-
-  let shopifyFlashListVersion = hasReactNativeWorkspace
-    ? findShopifyFlashListVersion(directory, packageJson)
+  const expoVersion = hasReactNativeWorkspace
+    ? resolveCatalogBackedDependencyVersion({
+        rootDirectory: directory,
+        rootPackageJson: packageJson,
+        packageName: "expo",
+        version: findExpoVersion(directory, packageJson),
+      })
     : null;
-  if (shopifyFlashListVersion !== null && isCatalogReference(shopifyFlashListVersion)) {
-    const catalogName = extractCatalogName(shopifyFlashListVersion);
-    let resolvedFlashListVersion = resolveCatalogVersion(
-      packageJson,
-      SHOPIFY_FLASH_LIST_PACKAGE_NAME,
-      directory,
-      catalogName,
-    );
-    if (!resolvedFlashListVersion) {
-      const monorepoRoot = findMonorepoRoot(directory);
-      if (monorepoRoot) {
-        const monorepoPackageJsonPath = path.join(monorepoRoot, "package.json");
-        if (isFile(monorepoPackageJsonPath)) {
-          resolvedFlashListVersion = resolveCatalogVersion(
-            readPackageJson(monorepoPackageJsonPath),
-            SHOPIFY_FLASH_LIST_PACKAGE_NAME,
-            monorepoRoot,
-            catalogName,
-          );
-        }
-      }
-    }
-    shopifyFlashListVersion = resolvedFlashListVersion ?? shopifyFlashListVersion;
-  }
+
+  const shopifyFlashListVersion = hasReactNativeWorkspace
+    ? resolveCatalogBackedDependencyVersion({
+        rootDirectory: directory,
+        rootPackageJson: packageJson,
+        packageName: SHOPIFY_FLASH_LIST_PACKAGE_NAME,
+        version: findShopifyFlashListVersion(directory, packageJson),
+      })
+    : null;
 
   // Only walk for reanimated once we already know it's an RN project —
   // reanimated implies React Native, so a web project can never declare
