+++ packages/oxlint-plugin-react-doctor/src/plugin/rules/react-native/rn-prefer-content-inset-adjustment.ts
@@ -1,14 +1,11 @@
-import { defineRule } from "../../utils/define-rule.js";
-import type { Rule } from "../../utils/rule.js";
+import { defineRetiredRule } from "../../utils/define-retired-rule.js";
 
-export const rnPreferContentInsetAdjustment = defineRule<Rule>({
+export const rnPreferContentInsetAdjustment = defineRetiredRule({
   id: "rn-prefer-content-inset-adjustment",
   title: "Manual safe-area inset adjustment",
   tags: ["test-noise"],
   requires: ["react-native"],
-  defaultEnabled: false,
   severity: "warn",
   recommendation:
     "Prefer native content inset adjustment only when it replaces manual inset plumbing; SafeAreaView wrappers are valid and intentionally ignored.",
-  create: () => ({}),
 });
